Created hole
So I was able to create a "working" version of a hole. Basically it first checks if the ball is within the radius of the hole. It then checks if the ball is in the radius for a given time. That way the ball doesn't go in the hole if it moves over it very fast. Couldn't find a tutorial or something online to help me with this so I had to chatGPT it. Will look to see how this could be improved